Adiciona exercício finalizado
